+++ src/drw_objects.cpp
@@ -454,7 +454,7 @@ void DRW_LType::parseCode(int code, dxfReader *reader){
 /*TODO: control max length permited */
 void DRW_LType::update(){
     double d =0;
-    size = path.size();
+    size = (int)path.size();
     for (int i = 0;  i< size; i++){
         d += fabs(path.at(i));
     }
